METRON-2078 Remove Storm dependency from metron-writer (merrimanr) closes apache/metron#1386
