+++ metron-analytics/metron-profiler-storm/pom.xml
@@ -84,7 +84,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.metron</groupId>
-            <artifactId>metron-writer</artifactId>
+            <artifactId>metron-writer-storm</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
