+++ metron-platform/metron-writer/metron-writer-common/src/main/java/org/apache/metron/writer/BatchSizePolicy.java
