+++ metron-platform/metron-writer/metron-writer-common/src/main/java/org/apache/metron/writer/kafka/KafkaWriter.java
