+++ metron-platform/metron-writer/metron-writer-common/src/main/resources/META-INF/NOTICE
