+++ metron-platform/metron-writer/metron-writer-common/src/test/java/org/apache/metron/writer/BatchSizePolicyTest.java
