+++ metron-platform/metron-writer/metron-writer-common/src/test/resources/log4j.properties
