+++ metron-platform/metron-writer/metron-writer-storm/src/main/java/org/apache/metron/writer/hdfs/ClonedSyncPolicyCreator.java
