+++ metron-platform/metron-writer/metron-writer-storm/src/test/java/org/apache/metron/writer/hdfs/ClonedSyncPolicyCreatorTest.java
